lesson 15 kadai

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,33 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // Task一覧を取得
-        $tasks = Task::all();
+        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+        }
+       
+        // Task一覧を取得
+        /*    $tasks = Task::all();
+            
+    
+            // メッセージ一覧ビューでそれを表示
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        */
+        // Welcomeビューでそれらを表示
+       return view('welcome', $data);
     }
 
     /**
@@ -54,14 +74,26 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
+         
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        
+        $request->user()->tasks()->create([
+        'status' => $request->status,
+        'content' => $request->content,
+        ]);
+        
+        return redirect('/');
+        /*
         // Task を作成
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
+        
         $task->save();
 
         // トップページへリダイレクトさせる
         return redirect('/');
+        */
     }
 
     /**
@@ -74,11 +106,19 @@ class TasksController extends Controller
     {
         // idの値でTaskを検索して取得
         $task = Task::findOrFail($id);
-
-        // Task詳細ビューでそれを表示
-        return view('tasks.show', [
+        
+         if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
             'task' => $task,
         ]);
+        
+        }
+         return redirect('/');
+
+        // Task詳細ビューでそれを表示
+       // return view('tasks.show', [
+        //    'task' => $task,
+       // ]);
     }
 
     /**
@@ -91,11 +131,19 @@ class TasksController extends Controller
     {
         // idの値でTaskを検索して取得
         $task = Task::findOrFail($id);
-
-        // Task編集ビューでそれを表示
-        return view('tasks.edit', [
+         
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
             'task' => $task,
         ]);
+        
+        }
+         return redirect('/');
+
+        // Task編集ビューでそれを表示
+        //return view('tasks.edit', [
+        //    'task' => $task,
+        //]);
     }
 
     /**
@@ -122,6 +170,7 @@ class TasksController extends Controller
 
         // トップページへリダイレクトさせる
         return redirect('/');
+        
     }
 
     /**
@@ -134,8 +183,14 @@ class TasksController extends Controller
     {
         // idの値でTaskを検索して取得
         $task = Task::findOrFail($id);
+        
+         // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         // Taskを削除
-        $task->delete();
+        //$task->delete();
 
         // トップページへリダイレクトさせる
         return redirect('/');
